Register address (#180)
* Added address part

Changed the register page to have the address fields. Need to update the controller to save the address.

* Ability to save address

Added in the RegisteredUserContoroller to be able to save the address in the address table. Only need to save the ID of the address in the user table

* Fixed address_id saving

Changed the way the user is saved in the RegisteredUserController

* Formatted code

* Check to see if address already exists

- Added a check to see if the address already exists

* Update RegisteredUserController.php